refactor: Restructure submodules in `evaluation`
