+++ haystack_experimental/evaluation/util/__init__.py
@@ -1,7 +1,3 @@
 # SPDX-FileCopyrightText: 2022-present deepset GmbH <info@deepset.ai>
 #
 # SPDX-License-Identifier: Apache-2.0
-
-from .harness import EvalRunOverrides, EvaluationHarness
-
-_all_ = ["EvaluationHarness", "EvalRunOverrides"]
